refactor: move pages in router
