+++ app/layout.tsx
@@ -1,7 +1,8 @@
 import type { ReactNode } from 'react';
-import { Layout } from '../src/pages/Layout/Layout';
 import App from '../src/app/App';
 import './globals.css';
+import { Flyout, ThemeToggle } from '@components';
+import { Navigation } from '@shared';
 
 export default function RootLayout({ children }: { children: ReactNode }) {
   return (
@@ -11,7 +12,17 @@ export default function RootLayout({ children }: { children: ReactNode }) {
       </head>
       <body>
         <App>
-          <Layout>{children}</Layout>
+          <div
+            className="box-border relative flex flex-col h-screen max-w px-6 py-3 dark:bg-neutral-900"
+            data-testid="layout"
+          >
+            <header className="flex justify-between">
+              <Navigation />
+              <ThemeToggle />
+            </header>
+            <main>{children}</main>
+            <Flyout />
+          </div>
         </App>
       </body>
     </html>
