+++ app/not-found.tsx
@@ -1,5 +1,15 @@
-import { NotFound } from '../src/pages';
-
 export default function NotFoundPage() {
-  return <NotFound />;
+  return (
+    <main
+      className="flex flex-col items-center justify-center min-h-[80vh]"
+      data-testid="not-found"
+    >
+      <h1 className="text-9xl font-extrabold text-rose-700/80 dark:text-rose-400 mb-6 tracking-wide">
+        404
+      </h1>
+      <span className="text-2xl font-semibold text-rose-900 dark:text-rose-300 mb-3">
+        Page Not Found
+      </span>
+    </main>
+  );
 }
