+++ src/app/App.tsx
@@ -1,5 +1,5 @@
 import { ErrorBoundary } from '@components';
-import { ThemeProvider } from './providers';
+import { StoreProvider, ThemeProvider } from './providers';
 import type { ReactNode } from 'react';
 
 type Props = {
@@ -9,9 +9,9 @@ type Props = {
 const App = ({ children }: Props) => {
   return (
     <ErrorBoundary>
-      {/* <StoreProvider> */}
-      <ThemeProvider>{children}</ThemeProvider>
-      {/* </StoreProvider> */}
+      <StoreProvider>
+        <ThemeProvider>{children}</ThemeProvider>
+      </StoreProvider>
     </ErrorBoundary>
   );
 };
