--- src/pages/About/About.test.tsx
@@ -1,19 +1,0 @@
-import { expect, it } from 'vitest';
-import '@testing-library/jest-dom';
-import { render, screen } from '@testing-library/react';
-import { About } from './About';
-import { createRoutesStub } from 'react-router';
-
-it('About page renders without error', () => {
-  const Stub = createRoutesStub([
-    {
-      path: '/about',
-      Component: About,
-    },
-  ]);
-
-  render(<Stub initialEntries={['/about']} />);
-
-  const aboutPage = screen.getByTestId('about-page');
-  expect(aboutPage).toBeInTheDocument();
-});
