--- src/pages/About/index.ts
@@ -1,1 +1,0 @@
-export { About } from './About';
