--- src/pages/Layout/index.ts
@@ -1,1 +1,0 @@
-export { Layout } from './Layout';
