--- src/pages/NotFound/NotFound.tsx
@@ -1,15 +1,0 @@
-export const NotFound = () => {
-  return (
-    <main
-      className="flex flex-col items-center justify-center min-h-[80vh]"
-      data-testid="not-found"
-    >
-      <h1 className="text-9xl font-extrabold text-rose-700/80 dark:text-rose-400 mb-6 tracking-wide">
-        404
-      </h1>
-      <span className="text-2xl font-semibold text-rose-900 dark:text-rose-300 mb-3">
-        Page Not Found
-      </span>
-    </main>
-  );
-};
