--- src/pages/NotFound/index.ts
@@ -1,1 +1,0 @@
-export { NotFound } from './NotFound';
