+++ src/pages/index.ts
@@ -1,4 +1,1 @@
 export { Main } from './Main';
-export { NotFound } from './NotFound';
-export { About } from './About';
-export { Layout } from './Layout';
