--- src/widget/index.ts
@@ -1,1 +1,0 @@
-export { CharacterWidget } from './CharacterWidget';
